+++ experiment_utils/conf/data/imagenette.yaml
@@ -1,7 +1,5 @@
 defaults:
   - /dataloader/base_dl@_global_.data._group_
-  - item_tfms: RandomResizedCrop_FlipItem
-  - batch_tfms: empty_list
 data_dir: /Data/imagenette2
 num_classes: 10
 size: 192
