--- experiment_utils/conf/data/item_tfms/RandomResizedCrop_FlipItem.yaml
@@ -1,5 +1,0 @@
-_target_: experiment_utils.utils.hydra_utils.load_args_list
-defaults:
-    # - empty
-  - /data/transforms/RandomResizedCrop@RandomResizedCrop
-  - /data/transforms/FlipItem@FlipItem
