--- experiment_utils/conf/data/item_tfms/Resize_FlipItem.yaml
@@ -1,5 +1,0 @@
-_target_: experiment_utils.utils.hydra_utils.load_args_list
-defaults:
-    # - empty
-  - /data/transforms/Resize@Resize
-  - /data/transforms/FlipItem@FlipItem
